+++ test/tasks/ods-build-go_test.go
@@ -5,6 +5,7 @@ import (
 	"os"
 	"path/filepath"
 	"runtime"
+	"strings"
 	"testing"
 
 	"github.com/opendevstack/pipeline/internal/command"
@@ -56,6 +57,11 @@ func TestTaskODSBuildGo(t *testing.T) {
 
 					sonarProject := sonar.ProjectKey(ctxt.ODS, "")
 					checkSonarQualityGate(t, ctxt.Clients.KubernetesClientSet, ctxt.Namespace, sonarProject, true, "OK")
+
+					wantLogMsg := "No sonar-project.properties present, using default:"
+					if !strings.Contains(string(ctxt.CollectedLogs), wantLogMsg) {
+						t.Fatalf("Want:\n%s\n\nGot:\n%s", wantLogMsg, string(ctxt.CollectedLogs))
+					}
 
 					b, _, err := command.Run(wsDir+"/docker/app", []string{})
 					if err != nil {
@@ -204,6 +210,12 @@ func TestTaskODSBuildGo(t *testing.T) {
 					}
 				},
 				WantRunSuccess: true,
+				PostRunFunc: func(t *testing.T, ctxt *tasktesting.TaskRunContext) {
+					notWantLogMsg := "No sonar-project.properties present, using default:"
+					if strings.Contains(string(ctxt.CollectedLogs), notWantLogMsg) {
+						t.Fatalf("Did not want:\n%s\n\nGot:\n%s", notWantLogMsg, string(ctxt.CollectedLogs))
+					}
+				},
 			},
 		})
 }